+++ src/Shared/Extensions/ApplicationInsightsServiceOptions.cs
@@ -90,6 +90,11 @@
         public string InstrumentationKey { get; set; }
 
         /// <summary>
+        /// Gets or sets the connection string for the application.
+        /// </summary>
+        public string ConnectionString { get; set; }
+
+        /// <summary>
         /// Gets or sets the application version reported with telemetries.
         /// </summary>
         public string ApplicationVersion { get; set; }
@@ -143,6 +148,10 @@
         /// </summary>
         public DependencyCollectionOptions DependencyCollectionOptions { get; }
 
+        /// <summary>
+        /// Copy the properties from this <see cref="ApplicationInsightsServiceOptions"/> to a target instance.
+        /// </summary>
+        /// <param name="target">Target instance to copy properties to.</param>
         internal void CopyPropertiesTo(ApplicationInsightsServiceOptions target)
         {
             if (this.DeveloperMode != null)
@@ -160,6 +169,7 @@
                 target.InstrumentationKey = this.InstrumentationKey;
             }
 
+            target.ConnectionString = this.ConnectionString;
             target.ApplicationVersion = this.ApplicationVersion;
             target.EnableAdaptiveSampling = this.EnableAdaptiveSampling;
             target.EnableDebugLogger = this.EnableDebugLogger;